Conflicts:
	src/docs/website/documentation/changes/dev.html
	src/ksp/ksp/examples/tutorials/ex59.c
	src/ksp/pc/impls/bddc/bddc.c
	src/ksp/pc/impls/bddc/bddcfetidp.c
	src/ksp/pc/impls/bddc/bddcprivate.c
	src/ksp/pc/impls/bddc/bddcschurs.c
	src/mat/impls/is/matis.c
	src/mat/impls/is/matis.h

+++ src/ts/impls/arkimex/arkimex.c
@@ -1216,7 +1216,7 @@ static PetscErrorCode PetscFormatRealArray(char buf[],size_t len,const char *fmt
 
   PetscFunctionBegin;
   for (i=0,p=buf,left=len; i<n; i++) {
-    ierr = PetscSNPrintfCount(p,left,fmt,&count,x[i]);CHKERRQ(ierr);
+    ierr = PetscSNPrintfCount(p,left,fmt,&count,(double)x[i]);CHKERRQ(ierr);
     if (count >= left) SETERRQ(PETSC_COMM_SELF,PETSC_ERR_ARG_OUTOFRANGE,"Insufficient space in buffer");
     left -= count;
     p    += count;
